Improve JAX example
