+++ examples/quickstart_jax/server.py
@@ -1,5 +1,3 @@
-"""Flower server example."""
-
 import flwr as fl
 
 if __name__ == "__main__":
